refactor: make AWS retry policy inheritable
