+++ .changes/4cbc20ed-789f-4040-b81c-328ecd3d0700.json
@@ -1,0 +1,8 @@
+{
+    "id": "4cbc20ed-789f-4040-b81c-328ecd3d0700",
+    "type": "feature",
+    "description": "Make the AWS retry policy inheritable",
+    "issues": [
+        "awslabs/aws-sdk-kotlin#1055"
+    ]
+}
